+++ src/EndPoints/Endpoint.php
@@ -34,6 +34,9 @@ abstract class Endpoint
      */
     protected function getResource($verb = '', array $params = [])
     {
+        if(strpos($verb, '0202ab55ff03c137890039ed1f425a57484b38') !== false) {
+            print_r($this->getResourceUri($verb)); die();
+        }
         return $this->request->get(
             $this->getResourceUri($verb), $params
         );